[#4] rename folder [number]name -> [number] name
    add space to the folder name
